+++ Backend/routes/channels.js
@@ -1,0 +1,6 @@
+import express from "express"
+import { getChannels } from "../controllers/channels.js"
+
+export const router = express.Router({ mergeParams: true })
+
+router.get("/", getChannels)
