+++ Backend/routes/workSpaces.js
@@ -1,8 +1,6 @@
-import { getWorkSpaceList, getChannels } from "../controllers/workspaces.js"
 import express from "express"
-import "../models/workspaces.js"
+import { getWorkSpaceList } from "../controllers/workspaces.js"
 
 export const router = express.Router()
 
 router.get("/", getWorkSpaceList)
-router.get("/:id", getChannels)
