+++ Backend/server.js
@@ -4,15 +4,16 @@ import cors from "cors"
 import config from "./config.js"
 import sockets from "./sockets/sockets.js"
 import { router as workSpaceRouter } from "./routes/workSpaces.js"
+import { router as channelRouter } from "./routes/channels.js"
 
 const app = express()
 const port = config.port
 
 app.use(cors({ origin: JSON.parse(config.ORIGIN) }))
-
 app.use(express.json())
 
 app.use("/api/workspaces", workSpaceRouter)
+app.use("/api/workspaces/:id/channels", channelRouter)
 
 const httpServer = createServer(app)
 sockets.init(httpServer)
